pyproject.toml
[project]
name = "music-tagger"
version = "0.1.0"
description = "自动化音乐元数据整理工具 - 监控NAS临时目录，自动补全标签/封面/歌词"
requires-python = ">=3.10"

dependencies = [
    "mutagen>=1.47",
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "thefuzz[speedup]>=0.22",
]

[project.optional-dependencies]
acoustid = [
    "pyacoustid>=1.3",
]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
music-tagger = "music_tagger.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
